+++ app/controllers/snippets-management/index.js
@@ -17,19 +17,6 @@ export default class SnippetsManagementIndexController extends Controller {
   @tracked isRemoveModalOpen = false;
   @tracked deletingSnippetList;
 
-  createSnippetList = task(async () => {
-    const administrativeUnit = await this.store.findRecord(
-      'administrative-unit',
-      this.currentSession.group.id
-    );
-    const snippetList = this.store.createRecord('snippet-list', {
-      createdOn: new Date(),
-      publisher: administrativeUnit,
-    });
-    await snippetList.save();
-    this.router.transitionTo('snippets-management.edit', snippetList);
-  });
-
   removeSnippetList = task(async () => {
     this.deletingSnippetList.deleteRecord();
     await this.deletingSnippetList.save();
