+++ app/router.js
@@ -27,6 +27,7 @@ Router.map(function () {
     this.route('callback');
   });
   this.route('snippets-management', function () {
+    this.route('new');
     this.route('edit', { path: '/edit/:id' }, function () {
       this.route('edit-snippet', { path: '/edit-snippet/:snippet_id' });
     });
